+++ pkg/analyzer/CHANGELOG.md
@@ -34,6 +34,7 @@
 * Remove deprecated `source_resource.dart`.
 * Remove `MultiplyInheritedExecutableElement`.
 * Remove `AnalyzeFunctionBodiesPredicate`.
+* Remove `normalParameterNames` and `optionalParameterNames` from `FunctionType`.
 
 ## 6.11.0
 * Un-deprecated `LibraryElement.exportedLibraries`.